Add WatchResourcesClient
